+++ azure-storage-file/azure/storage/file/fileservice.py
@@ -74,6 +74,7 @@ from ._deserialization import (
     _convert_xml_to_share_stats,
     _parse_file,
     _parse_share,
+    _parse_snapshot_share,
     _parse_directory,
 )
 from ._download_chunking import _download_file_chunks
@@ -529,8 +530,8 @@ class FileService(StorageClient):
 
         return self._perform_request(request, _convert_xml_to_service_properties)
 
-    def list_shares(self, prefix=None, marker=None, num_results=None,
-                    include_metadata=False, timeout=None):
+    def list_shares(self, prefix=None, marker=None, num_results=None, 
+                    include_metadata=False, timeout=None, include_snapshots=False):
         '''
         Returns a generator to list the shares under the specified account.
         The generator will lazily follow the continuation tokens returned by
@@ -557,8 +558,15 @@ class FileService(StorageClient):
             where the previous generator stopped.
         :param int timeout:
             The timeout parameter is expressed in seconds.
+        :param bool include_snapshots:
+            Specifies that share snapshots be returned in the response.
         '''
-        include = 'metadata' if include_metadata else None
+        include = 'snapshots' if include_snapshots else None
+        if include_metadata:
+            if include is not None:
+                include = include + ',metadata'
+            else:
+                include = 'metadata'
         operation_context = _OperationContext(location_lock=True)
         kwargs = {'prefix': prefix, 'marker': marker, 'max_results': num_results,
                   'include': include, 'timeout': timeout, '_context': operation_context}
@@ -586,9 +594,10 @@ class FileService(StorageClient):
             request may return up to 1000 shares and potentially a continuation
             token which should be followed to get additional resutls.
         :param string include:
-            Include this parameter to specify that the share's
-            metadata be returned as part of the response body. set this
-            parameter to string 'metadata' to get share's metadata.
+            Include this parameter to specify that either the share's
+            metadata, snapshots or both be returned as part of the response body. set this
+            parameter to string 'metadata' to get share's metadata. set this parameter to 'snapshots'
+            to get all the share snapshots. for both use 'snapshots,metadata'.
         :param int timeout:
             The timeout parameter is expressed in seconds.
         '''
@@ -657,7 +666,42 @@ class FileService(StorageClient):
             self._perform_request(request)
             return True
 
-    def get_share_properties(self, share_name, timeout=None):
+    def snapshot_share(self, share_name, metadata=None, quota=None, timeout=None):
+        '''
+        Creates a snapshot of an existing share under the specified account.
+
+        :param str share_name:
+            The name of the share to create a snapshot of.
+        :param metadata:
+            A dict with name_value pairs to associate with the
+            share as metadata. Example:{'Category':'test'}
+        :type metadata: a dict of str to str:
+        :param int quota:
+            Specifies the maximum size of the share, in gigabytes. Must be
+            greater than 0, and less than or equal to 5TB (5120).
+        :param int timeout:
+            The timeout parameter is expressed in seconds.
+        :return: snapshot properties
+        :rtype: azure.storage.file.models.Share
+        '''
+        _validate_not_none('share_name', share_name)
+        request = HTTPRequest()
+        request.method = 'PUT'
+        request.host_locations = self._get_host_locations()
+        request.path = _get_path(share_name)
+        request.query = {
+            'restype': 'share',
+            'comp':  'snapshot',
+            'timeout': _int_to_str(timeout),
+        }
+        request.headers = {
+            'x-ms-share-quota': _int_to_str(quota)
+        }
+        _add_metadata_headers(metadata, request)
+
+        return self._perform_request(request, _parse_snapshot_share, [share_name])
+
+    def get_share_properties(self, share_name, timeout=None, snapshot=None):
         '''
         Returns all user-defined metadata and system properties for the
         specified share. The data returned does not include the shares's
@@ -676,8 +720,9 @@ class FileService(StorageClient):
         request.host_locations = self._get_host_locations()
         request.path = _get_path(share_name)
         request.query = {
-            'restype': 'share',
-            'timeout': _int_to_str(timeout),
+             'restype': 'share',
+             'timeout': _int_to_str(timeout),
+             'sharesnapshot': _to_str(snapshot)
         }
 
         return self._perform_request(request, _parse_share, [share_name])
@@ -849,7 +894,7 @@ class FileService(StorageClient):
 
         return self._perform_request(request, _convert_xml_to_share_stats)
 
-    def delete_share(self, share_name, fail_not_exist=False, timeout=None):
+    def delete_share(self, share_name, fail_not_exist=False, timeout=None, snapshot=None, delete_snapshots=None):
         '''
         Marks the specified share for deletion. If the share
         does not exist, the operation fails on the service. By 
@@ -863,6 +908,10 @@ class FileService(StorageClient):
             exist. False by default.
         :param int timeout:
             The timeout parameter is expressed in seconds.
+       :param str snapshot:
+            A string that represents the snapshot version to delete, if applicable.
+        :type delete_snapshots:
+            One of the values listed in the :class:`~azure.storage.file.models.DeleteSnapshot` enum.
         :return: True if share is deleted, False share doesn't exist.
         :rtype: bool
         '''
@@ -871,9 +920,13 @@ class FileService(StorageClient):
         request.method = 'DELETE'
         request.host_locations = self._get_host_locations()
         request.path = _get_path(share_name)
+        request.headers = {
+            'x-ms-delete-snapshots': _to_str(delete_snapshots)
+        }
         request.query = {
-            'restype': 'share',
-            'timeout': _int_to_str(timeout),
+             'restype': 'share',
+             'timeout': _int_to_str(timeout),
+             'snapshot': _to_str(snapshot)
         }
 
         if not fail_not_exist:
@@ -981,7 +1034,7 @@ class FileService(StorageClient):
             self._perform_request(request)
             return True
 
-    def get_directory_properties(self, share_name, directory_name, timeout=None):
+    def get_directory_properties(self, share_name, directory_name, timeout=None, snapshot=None):
         '''
         Returns all user-defined metadata and system properties for the
         specified directory. The data returned does not include the directory's
@@ -994,6 +1047,8 @@ class FileService(StorageClient):
         :param int timeout:
             The timeout parameter is expressed in seconds.
         :return: properties for the specified directory within a directory object.
+        :param str snapshot:
+            A string that represents the snapshot version, if applicable.
         :rtype: :class:`~azure.storage.file.models.Directory`
         '''
         _validate_not_none('share_name', share_name)
@@ -1003,13 +1058,14 @@ class FileService(StorageClient):
         request.host_locations = self._get_host_locations()
         request.path = _get_path(share_name, directory_name)
         request.query = {
-            'restype': 'directory',
-            'timeout': _int_to_str(timeout),
+             'restype': 'directory',
+             'timeout': _int_to_str(timeout),
+             'sharesnapshot': _to_str(snapshot)
         }
 
         return self._perform_request(request, _parse_directory, [directory_name])
 
-    def get_directory_metadata(self, share_name, directory_name, timeout=None):
+    def get_directory_metadata(self, share_name, directory_name, timeout=None, snapshot=None):
         '''
         Returns all user-defined metadata for the specified directory.
 
@@ -1019,6 +1075,8 @@ class FileService(StorageClient):
             The path to the directory.
         :param int timeout:
             The timeout parameter is expressed in seconds.
+        :param str snapshot:
+            A string that represents the snapshot version, if applicable.
         :return:
             A dictionary representing the directory metadata name, value pairs.
         :rtype: dict(str, str)
@@ -1030,9 +1088,10 @@ class FileService(StorageClient):
         request.host_locations = self._get_host_locations()
         request.path = _get_path(share_name, directory_name)
         request.query = {
-            'restype': 'directory',
-            'comp': 'metadata',
-            'timeout': _int_to_str(timeout),
+             'restype': 'directory',
+             'comp': 'metadata',
+             'timeout': _int_to_str(timeout),
+             'sharesnapshot': _to_str(snapshot)
         }
 
         return self._perform_request(request, _parse_metadata)
@@ -1072,7 +1131,7 @@ class FileService(StorageClient):
 
     def list_directories_and_files(self, share_name, directory_name=None,
                                    num_results=None, marker=None, timeout=None,
-                                   prefix=None):
+                                   prefix=None, snapshot=None):
 
         '''
         Returns a generator to list the directories and files under the specified share.
@@ -1105,20 +1164,21 @@ class FileService(StorageClient):
             The timeout parameter is expressed in seconds.
         :param str prefix:
             List only the files and/or directories with the given prefix.
+        :param str snapshot:
+            A string that represents the snapshot version, if applicable.
         '''
         operation_context = _OperationContext(location_lock=True)
         args = (share_name, directory_name)
         kwargs = {'marker': marker, 'max_results': num_results, 'timeout': timeout,
-                  '_context': operation_context, 'prefix': prefix}
+                  '_context': operation_context, 'prefix': prefix, 'snapshot': snapshot}
 
         resp = self._list_directories_and_files(*args, **kwargs)
 
         return ListGenerator(resp, self._list_directories_and_files, args, kwargs)
 
     def _list_directories_and_files(self, share_name, directory_name=None,
-                                    marker=None, max_results=None, timeout=None,
-                                    prefix=None, _context=None):
-
+                                   marker=None, max_results=None, timeout=None,
+                                    prefix=None, _context=None, snapshot=None):
         '''
         Returns a list of the directories and files under the specified share.
 
@@ -1143,6 +1203,8 @@ class FileService(StorageClient):
             The timeout parameter is expressed in seconds.
         :param str prefix:
             List only the files and/or directories with the given prefix.
+        :param str snapshot:
+            A string that represents the snapshot version, if applicable.
         '''
         _validate_not_none('share_name', share_name)
         request = HTTPRequest()
@@ -1150,18 +1212,19 @@ class FileService(StorageClient):
         request.host_locations = self._get_host_locations()
         request.path = _get_path(share_name, directory_name)
         request.query = {
-            'restype': 'directory',
-            'comp': 'list',
-            'prefix': _to_str(prefix),
-            'marker': _to_str(marker),
-            'maxresults': _int_to_str(max_results),
-            'timeout': _int_to_str(timeout),
+             'restype': 'directory',
+             'comp': 'list',
+             'prefix': _to_str(prefix),
+             'marker': _to_str(marker),
+             'maxresults': _int_to_str(max_results),
+             'timeout': _int_to_str(timeout),
+             'sharesnapshot': _to_str(snapshot)
         }
 
         return self._perform_request(request, _convert_xml_to_directories_and_files,
                                      operation_context=_context)
 
-    def get_file_properties(self, share_name, directory_name, file_name, timeout=None):
+    def get_file_properties(self, share_name, directory_name, file_name, timeout=None, snapshot=None):
         '''
         Returns all user-defined metadata, standard HTTP properties, and
         system properties for the file. Returns an instance of :class:`~azure.storage.file.models.File` with
@@ -1175,6 +1238,8 @@ class FileService(StorageClient):
             Name of existing file.
         :param int timeout:
             The timeout parameter is expressed in seconds.
+        :param str snapshot:
+            A string that represents the snapshot version, if applicable.
         :return: a file object including properties and metadata.
         :rtype: :class:`~azure.storage.file.models.File`
         '''